Re-add fan of gales
Fan of gales is an evokable that is distinct from the others in that it
can reposition enemies, remove clouds, etc. Though can be said that it is
situational and has limited usage, it gives an unique way of resolving
situations and sacrificing it for the sake of freeing inventory slot
hardly seems like a reasonable trade.