Re-add targeting from Summon Lightning Spire
SLS shoots lightning bolt that bounces, which means its placement itself
has significant importance. For instance, lightning spire can be placed
right in front of corridor to maximize the number of enemy hit by its
bolt, or placed diagonally to allow double-zap.

If it is absolutely necessary to make optimal placement cost player
movement(and turns), it might be changed so that lightning spire can be
swapped with character. However I am not convienced that it is the case.

Yes, it outperforms other lv4 spells. Its damage can be adjusted.

Yes, instant barricade can be powerful. Its hp can be reduced.

No, randomizing its placement does not solve its power balance issue. It
just makes it more tedious and irritating.

And no, autotargeting is not superior to manual targeting. It can be
used as a mean to reduce redundant keypresses, but it should not come at
a cost of reducing stratagic depth.